+++ packages/core/src/types/foundry-v13-essentials.d.ts
@@ -187,6 +187,18 @@ interface ClientSettings {
   get(module: string, setting: string): any;
   set(module: string, setting: string, value: any): Promise<any>;
   register(module: string, setting: string, config: any): void;
+  registerMenu(
+    namespace: string,
+    key: string,
+    data: {
+      name: string;
+      label: string;
+      hint: string;
+      icon: string;
+      type: new () => foundry.applications.api.ApplicationV2; // Application class constructor
+      restricted: boolean;
+    }
+  ): void;
 }
 
 interface FoundryUser {