+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.24.1](https://github.com/rayners/fvtt-seasons-and-stars/compare/seasons-and-stars-v0.24.0...seasons-and-stars-v0.24.1) (2025-10-30)
+
+
+### Bug Fixes
+
+* replace calendar dropdown with menu button ([#448](https://github.com/rayners/fvtt-seasons-and-stars/issues/448)) ([80d2c64](https://github.com/rayners/fvtt-seasons-and-stars/commit/80d2c642fa1d339ca89564c1a2e92de3a7966f7d))
+
 ## [0.24.0](https://github.com/rayners/fvtt-seasons-and-stars/compare/seasons-and-stars-v0.23.1...seasons-and-stars-v0.24.0) (2025-10-28)